Andy Ross: Attached is my current a4-sound.xml file, if folks want to try it.  I've tweaked a bunch of things to make it more pleasing to my ears.  The engine pitch is scaled to the N1 RPM and is played at a roughly constant volume rather than dying off at high thrust, and the roar scales linearly with thrust rather than with the logarithm of thrust, which matches my experience on jetliners better (most of the sound exists only at the highest thrust settings).  At idle, all you hear is the engine whine -- no thrust rumble.  The gear lock sounds plays correctly now when the gear reach minimum extension.
I really like it -- you can definitely hear the engine spool up and
down now, and associate a given sound with a thrust setting.  My
approaches in the A-4 have actually gotten better because of it.  I'll
probably continue tweaking in the future, but it's good enough for
government work as-is, IMHO.